refactor: Floor를 만드는 정적팩토리메서드를 정적팩토리 클래스에게 이전
Floor의 주요 역할은 참가자의 위치를 변경시키는 것인데 현재 정적 팩토리 메서드 때문에 생성이 핵심인 것으로 오해의 소지가 있으므로 변경
