+++ src/test/java/domain/ladder/FloorTest.java
@@ -17,7 +17,7 @@ class FloorTest {
         BridgeGenerator bridgeGenerator = new PickedBridgeGenerator(List.of(false, true, false));
 
         //when
-        Floor floor = Floor.createByStrategy(bridgeGenerator, new Width(3));
+        Floor floor = FloorFactory.createByStrategy(bridgeGenerator, new Width(3));
 
         //then
         assertThat(floor.getBridges()).containsExactly(Bridge.NO_BRIDGE, Bridge.BRIDGE, Bridge.NO_BRIDGE);
@@ -28,7 +28,7 @@ class FloorTest {
     void moveAlongBridge() {
         //given
         BridgeGenerator bridgeGenerator = new PickedBridgeGenerator(List.of(true, false, true));
-        Floor floor = Floor.createByStrategy(bridgeGenerator, new Width(3));
+        Floor floor = FloorFactory.createByStrategy(bridgeGenerator, new Width(3));
 
         //when & then
         assertAll(
